% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve_cpp}
\alias{convolve_cpp}
\title{Convolution.}
\usage{
convolve_cpp(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector.}
}
\value{
The result as a numeric vector.
}
\description{
Example function taken from Rcpp vignette "Rcpp-introduction".
This function is callable from R as a normal R function, it is thus tested 
as a normal R function.
}
