% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{hmdbToSqlite}
\alias{hmdbToSqlite}
\title{Convert HMDB Metabolites to a SQLite database.}
\usage{
hmdbToSqlite(sqlite.file = "hmdb.sqlite")
}
\arguments{
\item{sqlite.file}{Path to the SQLite file to create.}
}
\value{
Nothing.
}
\description{
Create an SQLite compound database
}
\examples{
# Copy all HMDB entries into a new SQLite database:
biodbHmdb::hmdbToSqlite('myfile.sqlite')
}
